Feat: create MessageBar file
