+++ src/app/my-page/message/panel/MessageContainer.tsx
@@ -2,73 +2,14 @@
 
 import { useCallback, useEffect, useState } from 'react'
 import { AxiosResponse } from 'axios'
-import { Button, Stack, TextField, Typography } from '@mui/material'
+import { Stack, Typography } from '@mui/material'
 import useAxiosWithAuth from '@/api/config'
 import { IMessageListData } from '@/types/IMessage'
-import CuButton from '@/components/CuButton'
 import useSelectCheckBox from '@/hook/useSelectCheckbox'
 import useToast from '@/hook/useToast'
 import MessageList from './MessageList'
 import useMessageListState from '@/states/useMessageListState'
-
-interface ISearchBarProps {
-  setSearchKeyword: (keyword: string) => void
-  setIsManageMode: (isManageMode: boolean) => void
-  handleMessageSearch: () => void
-}
-
-const SearchBar = ({
-  setSearchKeyword,
-  setIsManageMode,
-  handleMessageSearch,
-}: ISearchBarProps) => {
-  return (
-    <Stack direction="row">
-      <TextField
-        placeholder="사람을 검색해 주세요."
-        onChange={(e) => setSearchKeyword(e.target.value)}
-      />
-      <Button variant="contained" onClick={handleMessageSearch}>
-        검색
-      </Button>
-      <CuButton
-        variant="text"
-        action={() => setIsManageMode(true)}
-        message="관리"
-      />
-    </Stack>
-  )
-}
-
-interface IManageBarProps {
-  isSelectAll: boolean
-  handleSelectAll: () => void
-  handleUnselectAll: () => void
-  handleDelete: () => void
-}
-
-const ManageBar = ({
-  isSelectAll,
-  handleSelectAll,
-  handleUnselectAll,
-  handleDelete,
-}: IManageBarProps) => {
-  return (
-    // 관리 모드 나가기 버튼?
-    <Stack direction="row">
-      {isSelectAll ? (
-        <CuButton
-          variant="text"
-          action={handleUnselectAll}
-          message="전체 선택 해제"
-        />
-      ) : (
-        <CuButton variant="text" action={handleSelectAll} message="전체 선택" />
-      )}
-      <CuButton variant="text" action={handleDelete} message="삭제" />
-    </Stack>
-  )
-}
+import { SearchBar, ManageBar } from './MessageBar'
 
 interface IMessageContainerProps {
   originalMessageData: IMessageListData[] | undefined
@@ -143,11 +84,11 @@ const MessageContainer = ({
       <Stack spacing={2}>
         {isManageMode ? (
           <ManageBar
-            isSelectAll={isSelectedAll(messageList)}
+            isSelectedAll={isSelectedAll(messageList)}
             handleSelectAll={() =>
               selectAll(messageList.map((message) => message.targetId))
             }
-            handleUnselectAll={() => unselectAll()}
+            handleUnselectAll={unselectAll}
             handleDelete={handleDelete}
           />
         ) : (
